+++ lib/output/constructor.go
@@ -81,6 +81,7 @@ type Config struct {
 	RedisPubSub   RedisPubSubConfig          `json:"redis_pubsub" yaml:"redis_pubsub"`
 	ScaleProto    ScaleProtoConfig           `json:"scalability_protocols" yaml:"scalability_protocols"`
 	STDOUT        STDOUTConfig               `json:"stdout" yaml:"stdout"`
+	Websocket     writer.WebsocketConfig     `json:"websocket" yaml:"websocket"`
 	ZMQ4          *writer.ZMQ4Config         `json:"zmq4,omitempty" yaml:"zmq4,omitempty"`
 	Processors    []processor.Config         `json:"processors" yaml:"processors"`
 }
@@ -108,6 +109,7 @@ func NewConfig() Config {
 		RedisPubSub:   NewRedisPubSubConfig(),
 		ScaleProto:    NewScaleProtoConfig(),
 		STDOUT:        NewSTDOUTConfig(),
+		Websocket:     writer.NewWebsocketConfig(),
 		ZMQ4:          writer.NewZMQ4Config(),
 		Processors:    []processor.Config{},
 	}